refactor: 영화 목록 스켈레톤 로딩 구현 로직 수정
