+++ index.html
@@ -13,7 +13,10 @@
   </head>
   <body>
     <div id="wrap">
-      <main class="container"></main>
+      <main class="container">
+        <h2 id="movieKindTitle"></h2>
+        <ul class="thumbnail-list"></ul>
+      </main>
     </div>
     <script src="src/main.ts" type="module"></script>
   </body>
