+++ src/view/moreButton.ts
@@ -1,14 +1,13 @@
 import Button from '../component/Button';
 import { addMovies } from '../domain/addMovies';
 import { $ } from '../util/selector';
-import { hideSkeletons, renderSkeletons } from './render/renderSkeletons';
+import { hideSkeletons, movieListSkeletons } from './render/skeleton/movieListSkeletons';
 
 const handleMoreButtonClick = async (page: number, totalPages: number, keyword?: string) => {
   const container = $('.thumbnail-list') as HTMLElement;
   if (!container) return;
 
-  const skeletonList = renderSkeletons();
-  container.appendChild(skeletonList);
+  movieListSkeletons();
 
   const movies = await addMovies(page, keyword);
   container.appendChild(movies);
